+++ packages/wonder-blocks-search-field/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-search-field
 
+## 5.1.58
+
+### Patch Changes
+
+- @khanacademy/wonder-blocks-form@7.5.1
+- @khanacademy/wonder-blocks-icon-button@11.1.1
+
+## 5.1.57
+
+### Patch Changes
+
+- Updated dependencies [12169b4]
+    - @khanacademy/wonder-blocks-form@7.5.0
+
 ## 5.1.56
 
 ### Patch Changes